Use component width for `RenderHtml` rather than screen width
The old implementation was pretty naive – using the width of the entire screen for all HTML content, subtracting the page margins for images. We had an image overlapping a timeline because that took extra horizontal space itself, which was unaccounted for. This update lets the component retrieve its own width and use that instead, which allows the component to be rendered anywhere on the page. We should note that this requires a width calculation for each instance of the `RenderHtml` component.