Inject query parameter into waste containers url
This ensures that the browser displays the correct fraction when opening such a link more than once. It seems that multiple links which only vary for the fragment are considered the same links, which confirms to the specification, resulting in the page not being refreshed after the first time and the incorrect fraction displaying.