Stop the prop spteading practice
We should stop this to prevent more occurrences of the unnoticed bug with an `injectedJavaScript` prop being spelled incorrectly fixed in the previous commit.
The `react/jsx-props-no-spreading` ESLint rule doesn’t give us an option to allow `{...props}` but disallow `{...{prop1, prop2, prop3}}` so we’ll just have to remember or document this.